requirements.txt
astroid==2.3.1
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
cloudinary==1.19.1
colorama==0.4.1
defusedxml==0.6.0
Django==2.2
django-allauth==0.40.0
django-countries==5.5
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-jsonfield==1.3.1
future==0.17.1
h11==0.8.1
h2==3.1.1
hpack==3.0.0
http3==0.6.7
hyperframe==5.2.0
idna==2.8
importlib-metadata==0.23
isort==4.3.21
jedi==0.14.1
lazy-object-proxy==1.4.2
mccabe==0.6.1
mock==4.0.1
more-itertools==7.2.0
oauthlib==3.1.0
Paperboy==1.0.1
parso==0.5.1
Pillow==6.2.0
pluggy==0.13.0
psycopg2==2.8.4
pycodestyle==2.5.0
pydocstyle==4.0.1
pyflakes==2.1.1
pylint==2.4.2
python-dateutil==2.8.1
python-decouple==3.1
python-jsonrpc-server==0.2.0
python-language-server==0.28.3
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-async==0.6.2
requests-oauthlib==1.2.0
rfc3986==1.3.2
rope==0.14.0
six==1.12.0
snowballstemmer==2.0.0
sqlparse==0.2.4
stripe==2.37.2
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
yapf==0.28.0
zipp==0.6.0
